+++ metadata.rb
@@ -3,7 +3,7 @@ maintainer_email "a.rosenhagen@ffuenf.de"
 license          "Apache 2.0"
 description      "installs/configures dop_base"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "0.0.5"
+version          "0.0.6"
 
 %w{ debian }.each do |os|
 	supports os